feat: use event name instead of const to be psr-14 compliant
