+++ src/EventDispatcher/GuzzleCacheErrorEvent.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace M6Web\Bundle\GuzzleHttpBundle\EventDispatcher;
 
 /**
- * Class GuzzleCacheEvent
+ * Class GuzzleCacheErrorEvent
  */
-class GuzzleCacheEvent extends AbstractGuzzleCacheEvent
+class GuzzleCacheErrorEvent extends AbstractGuzzleCacheEvent
 {
 }
